Split settings-ui.js into per-domain modules under ui/settings/
settings-ui.js: 754 -> 183 lines. Sub-dialog builders moved to
ui/settings/dialogs/{core-limits,collect-targets,storage-list,task-priority}.js;
ROW_BG/addRowSeparator/addStandardReset to ui/settings/helpers.js;
resetXxxSettings helpers to ui/settings/reset-helpers.js. Pref
registration, custom reset, and sub-dialog wiring stay in the entry
point so the SettingsTable build order is unchanged.
